requirements.txt
pymongo[srv]==4.6
python-dotenv
sentence-transformers
pandas